social_media_api/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
